pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "neolore-queue-analytics"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
  "duckdb>=1.2.0",
  "httpx>=0.27.0",
  "pandas>=2.2.0",
  "pyarrow>=15.0.0",
  "python-dotenv>=1.0.0",
  "tenacity>=8.2.0",
  "azure-identity>=1.16.0",
  "azure-storage-blob>=12.20.0"
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0.0"
]

[tool.setuptools.packages.find]
include = ["pipeline*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
